+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./Inputs";
 export * from "./Outputs";
 export * from "./Logo";
+export * from "./Piano";
